+++ library/model.class.php
@@ -7,7 +7,7 @@
  *
  * PHP version 5.4
  *
- * @author    Brock Hensley <Brock@GPFC.com>
+ * @author    Brock Hensley <Brock@brockhensley.com>
  *
  * @version   v00.00.0000
  *